+++ Classes/Service/UrlService.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace YoastSeoForTypo3\YoastSeo\Service;
 
+use Psr\EventDispatcher\EventDispatcherInterface;
 use Psr\Http\Message\UriInterface;
 use TYPO3\CMS\Backend\Routing\UriBuilder;
 use TYPO3\CMS\Backend\Utility\BackendUtility;
@@ -13,7 +14,7 @@ use TYPO3\CMS\Core\Routing\RouteNotFoundException;
 use TYPO3\CMS\Core\SingletonInterface;
 use TYPO3\CMS\Core\Site\Entity\Site;
 use TYPO3\CMS\Core\Site\SiteFinder;
-use TYPO3\CMS\Core\Utility\GeneralUtility;
+use YoastSeoForTypo3\YoastSeo\Event\ModifyPreviewUrlEvent;
 use YoastSeoForTypo3\YoastSeo\Utility\YoastUtility;
 
 class UrlService implements SingletonInterface
@@ -21,6 +22,8 @@ class UrlService implements SingletonInterface
     public function __construct(
         protected UriBuilder $uriBuilder,
         protected SiteFinder $siteFinder,
+        protected PageRepository $pageRepository,
+        protected EventDispatcherInterface $eventDispatcher,
     ) {}
 
     public function getPreviewUrl(
@@ -41,30 +44,29 @@ class UrlService implements SingletonInterface
         $rootLine = $this->getRootLine($pageId);
         $site = $this->getSite($pageId, $rootLine);
 
-        if ($site !== null) {
-            $uriToCheck = YoastUtility::fixAbsoluteUrl(
-                (string)$this->generateUri($site, $pageId, $languageId, $additionalGetVars)
-            );
-
-            foreach ($GLOBALS['TYPO3_CONF_VARS']['SC_OPTIONS'][self::class]['urlToCheck'] ?? [] as $_funcRef) {
-                $_params = [
-                    'urlToCheck' => $uriToCheck,
-                    'site' => $site,
-                    'finalPageIdToShow' => $pageId,
-                    'languageId' => $languageId,
-                ];
-
-                $uriToCheck = GeneralUtility::callUserFunction($_funcRef, $_params, $this);
-            }
-            return $uriToCheck;
+        if ($site === null) {
+            return '';
         }
-        return '';
+
+        $uriToCheck = YoastUtility::fixAbsoluteUrl(
+            (string)$this->generateUri($site, $pageId, $languageId, $additionalGetVars)
+        );
+
+        /** @var ModifyPreviewUrlEvent $event */
+        $event = $this->eventDispatcher->dispatch(
+            new ModifyPreviewUrlEvent(
+                $uriToCheck,
+                $site,
+                $pageId,
+                $languageId
+            )
+        );
+        return $event->getUrl();
     }
 
     public function checkMountPoint(int &$pageId, string &$additionalGetVars): void
     {
-        $pageRepository = GeneralUtility::makeInstance(PageRepository::class);
-        $mountPointInformation = $pageRepository->getMountPointInfo($pageId);
+        $mountPointInformation = $this->pageRepository->getMountPointInfo($pageId);
         if ($mountPointInformation && $mountPointInformation['overlay']) {
             // New page id
             $pageId = $mountPointInformation['mount_pid'];