+++ CHANGELOG.md
@@ -5,6 +5,12 @@ This changelog is according to [Keep a Changelog](http://keepachangelog.com).
 All notable changes to this project will be documented in this file.
 We will follow [Semantic Versioning](http://semver.org/).
 
+## UNRELEASED
+
+### Breaking
+- Replaced the `urlToCheck` with a new `ModifyPreviewUrl` event
+  - TODO: Add documentation 
+
 ## 10.0.0 October 15, 2024
 ### Breaking
 - Dropped support for TYPO3 10